Split declarations from implementations
