+++ lib/core/json_exception.h
@@ -1,5 +1,5 @@
-#ifndef JSON_EXCEPTION
-#define JSON_EXCEPTION
+#ifndef JSON_EXCEPTION_H
+#define JSON_EXCEPTION_H
 
 #include <string>
 
@@ -7,16 +7,16 @@
 enum ExceptionType {LEXER, SYNTAX, SEMANTIC};
 
 
-// specialized exception for json implementation
+// specialized exception for wjson implementation
 class JSONException : public std::exception
 {
  public:
 
   // construct a "normal" error exception
-  JSONException(ExceptionType type, const std::string& msg, int line, int column);
+  JSONException(ExceptionType, const std::string&, int, int);
 
   // construct an error exception without a line and column
-  JSONException(ExceptionType type, const std::string& msg);
+  JSONException(ExceptionType, const std::string&);
 
   // return a string representation for printing
   std::string to_string() const;
@@ -32,32 +32,4 @@ class JSONException : public std::exception
 };
 
 
-JSONException::JSONException(ExceptionType t, const std::string& m, int l, int c)
-  : type(t), message(m), line(l), column(c), has_line_column(true)
-{
-}
-
-
-JSONException::JSONException(ExceptionType t, const std::string& m)
-  : type(t), message(m), has_line_column(false)
-{
-}
-
-
-std::string JSONException::to_string() const
-{
-  std::string s;
-  switch(type) {
-    case LEXER: s = "Lexer"; break;
-    case SYNTAX: s = "Parser"; break;
-    case SEMANTIC: s = "Type"; break;
-  }
-  s += " Error: " + message;
-  if (has_line_column)
-    s += " at line " + std::to_string(line) +
-      " column " + std::to_string(column);
-  return s;
-}
-
-
-#endif
+#endif // ifndef JSON_EXCEPTION_H
